move TemplateWatcher functionality directly into ManagedContext
